spl_token_2022_interface::extension

Module confidential_transfer

Confidential Transfer extension

Modules

instruction

Confidential Transfer Extension instructions

Structs

ConfidentialTransferAccount

Confidential account state

ConfidentialTransferMint

Confidential transfer mint configuration

Constants

DEFAULT_MAXIMUM_PENDING_BALANCE_CREDIT_COUNTER

The default maximum pending balance credit counter.

MAXIMUM_DEPOSIT_TRANSFER_AMOUNT

Maximum bit length of any deposit or transfer amount

PENDING_BALANCE_LO_BIT_LENGTH

Bit length of the low bits of pending balance plaintext

Type Aliases

DecryptableBalance

Authenticated encryption containing an account balance

EncryptedBalance

ElGamal ciphertext containing an account balance