spl_token_2022_interface::extension::confidential_transfer::instruction

Enum ProofInstruction

#[repr(u8)]
pub enum ProofInstruction {

    CloseContextState = 0,
    VerifyZeroCiphertext = 1,
    VerifyCiphertextCiphertextEquality = 2,
    VerifyCiphertextCommitmentEquality = 3,
    VerifyPubkeyValidity = 4,
    VerifyPercentageWithCap = 5,
    VerifyBatchedRangeProofU64 = 6,
    VerifyBatchedRangeProofU128 = 7,
    VerifyBatchedRangeProofU256 = 8,
    VerifyGroupedCiphertext2HandlesValidity = 9,
    VerifyBatchedGroupedCiphertext2HandlesValidity = 10,
    VerifyGroupedCiphertext3HandlesValidity = 11,
    VerifyBatchedGroupedCiphertext3HandlesValidity = 12,

}

Variants

CloseContextState = 0

Close a zero-knowledge proof context state.

Accounts expected by this instruction: 0. [writable] The proof context account to close

[writable] The destination account for lamports

[signer] The context account’s owner

Data expected by this instruction: None

VerifyZeroCiphertext = 1

Verify a zero-ciphertext proof.

A zero-ciphertext proof certifies that an ElGamal ciphertext encrypts the value zero.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. ZeroCiphertextProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyCiphertextCiphertextEquality = 2

Verify a ciphertext-ciphertext equality proof.

A ciphertext-ciphertext equality proof certifies that two ElGamal ciphertexts encrypt the same message.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. CiphertextCiphertextEqualityProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyCiphertextCommitmentEquality = 3

Verify a ciphertext-commitment equality proof.

A ciphertext-commitment equality proof certifies that an ElGamal ciphertext and a Pedersen commitment encrypt/encode the same message.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. CiphertextCommitmentEqualityProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyPubkeyValidity = 4

Verify a public key validity zero-knowledge proof.

A public key validity proof certifies that an ElGamal public key is well-formed and the prover knows the corresponding secret key.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. PubkeyValidityData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyPercentageWithCap = 5

Verify a percentage-with-cap proof.

A percentage-with-cap proof certifies that a tuple of Pedersen commitments satisfy a percentage relation.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. PercentageWithCapProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyBatchedRangeProofU64 = 6

Verify a 64-bit batched range proof.

A batched range proof is defined with respect to a sequence of Pedersen commitments [C_1, ..., C_N] and bit-lengths [n_1, ..., n_N]. It certifies that each commitment C_i is a commitment to a positive number of bit-length n_i. Batch verifying range proofs is more efficient than verifying independent range proofs on commitments C_1, ..., C_N separately.

The bit-length of a batched range proof specifies the sum of the individual bit-lengths n_1, ..., n_N. For example, this instruction can be used to certify that two commitments C_1 and C_2 each hold positive 32-bit numbers.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. BatchedRangeProofU64Data if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyBatchedRangeProofU128 = 7

Verify 128-bit batched range proof.

The bit-length of a batched range proof specifies the sum of the individual bit-lengths n_1, ..., n_N. For example, this instruction can be used to certify that two commitments C_1 and C_2 each hold positive 64-bit numbers.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. BatchedRangeProofU128Data if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyBatchedRangeProofU256 = 8

Verify 256-bit batched range proof.

The bit-length of a batched range proof specifies the sum of the individual bit-lengths n_1, ..., n_N. For example, this instruction can be used to certify that four commitments [C_1, C_2, C_3, C_4] each hold positive 64-bit numbers.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. BatchedRangeProofU256Data if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyGroupedCiphertext2HandlesValidity = 9

Verify a grouped-ciphertext with 2 handles validity proof.

A grouped-ciphertext validity proof certifies that a grouped ElGamal ciphertext is well-defined, i.e. the ciphertext can be decrypted by private keys associated with its decryption handles.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. GroupedCiphertext2HandlesValidityProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyBatchedGroupedCiphertext2HandlesValidity = 10

Verify a batched grouped-ciphertext with 2 handles validity proof.

A batched grouped-ciphertext validity proof certifies the validity of two grouped ElGamal ciphertext that are encrypted using the same set of ElGamal public keys. A batched grouped-ciphertext validity proof is shorter and more efficient than two individual grouped-ciphertext validity proofs.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. BatchedGroupedCiphertext2HandlesValidityProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyGroupedCiphertext3HandlesValidity = 11

Verify a grouped-ciphertext with 3 handles validity proof.

A grouped-ciphertext validity proof certifies that a grouped ElGamal ciphertext is well-defined, i.e. the ciphertext can be decrypted by private keys associated with its decryption handles.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. GroupedCiphertext3HandlesValidityProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

VerifyBatchedGroupedCiphertext3HandlesValidity = 12

Verify a batched grouped-ciphertext with 3 handles validity proof.

A batched grouped-ciphertext validity proof certifies the validity of two grouped ElGamal ciphertext that are encrypted using the same set of ElGamal public keys. A batched grouped-ciphertext validity proof is shorter and more efficient than two individual grouped-ciphertext validity proofs.

Accounts expected by this instruction:

There are four ways to structure the accounts, depending on whether the proof is provided as instruction data or in a separate account, and whether a proof context is created.

Proof in instruction data, no context state:

No accounts are required.

Proof in instruction data, with context state:

[writable] The proof context account to create.

[] The proof context account owner.

Proof in account, no context state:

[] Account to read the proof from.

Proof in account, with context state:

[] Account to read the proof from.

[writable] The proof context account to create.

[] The proof context account owner.

The instruction expects either: i. BatchedGroupedCiphertext3HandlesValidityProofData if proof is provided as instruction data ii. u32 byte offset if proof is provided as an account

Implementations

impl ProofInstruction

pub fn encode_verify_proof<T, U>( &self, context_state_info: Option<ContextStateInfo<'_>>, proof_data: &T, ) -> Instruction
where T: Pod + ZkProofData<U>, U: Pod,

pub fn encode_verify_proof_from_account( &self, context_state_info: Option<ContextStateInfo<'_>>, proof_account: &Address, offset: u32, ) -> Instruction

pub fn instruction_type(input: &[u8]) -> Option<ProofInstruction>

pub fn proof_data<T, U>(input: &[u8]) -> Option<&T>
where T: Pod + ZkProofData<U>, U: Pod,