spl_token_2022_interface::extension::confidential_transfer_fee::instruction

Enum ConfidentialTransferFeeInstruction

#[repr(u8)]
pub enum ConfidentialTransferFeeInstruction {
    InitializeConfidentialTransferFeeConfig = 0,
    WithdrawWithheldTokensFromMint = 1,
    WithdrawWithheldTokensFromAccounts = 2,
    HarvestWithheldTokensToMint = 3,
    EnableHarvestToMint = 4,
    DisableHarvestToMint = 5,
}

Confidential Transfer extension instructions

Variants

InitializeConfidentialTransferFeeConfig = 0

Initializes confidential transfer fees for a mint.

The ConfidentialTransferFeeInstruction::InitializeConfidentialTransferFeeConfig instruction requires no signers and MUST be included within the same Transaction as TokenInstruction::InitializeMint. Otherwise another party can initialize the configuration.

The instruction fails if the TokenInstruction::InitializeMint instruction has already executed for the mint.

Accounts expected by this instruction:

[writable] The SPL Token mint.

Data expected by this instruction: InitializeConfidentialTransferFeeConfigData

WithdrawWithheldTokensFromMint = 1

Transfer all withheld confidential tokens in the mint to an account. Signed by the mint’s withdraw withheld tokens authority.

The withheld confidential tokens are aggregated directly into the destination available balance.

In order for this instruction to be successfully processed, it must be accompanied by the VerifyCiphertextCiphertextEquality instruction of the zk_elgamal_proof program in the same transaction or the address of a context state account for the proof must be provided.

Accounts expected by this instruction:

Single owner/delegate

[writable] The token mint. Must include the TransferFeeConfig extension.

[writable] The fee receiver account. Must include the TransferFeeAmount and ConfidentialTransferAccount extensions.

[] Instructions sysvar if VerifyCiphertextCiphertextEquality is included in the same transaction or context state account if VerifyCiphertextCiphertextEquality is pre-verified into a context state account.

[signer] The mint’s withdraw_withheld_authority.

Multisignature owner/delegate

[writable] The token mint. Must include the TransferFeeConfig extension.

[writable] The fee receiver account. Must include the TransferFeeAmount and ConfidentialTransferAccount extensions.

[] Instructions sysvar if VerifyCiphertextCiphertextEquality is included in the same transaction or context state account if VerifyCiphertextCiphertextEquality is pre-verified into a context state account.

[] The mint’s multisig withdraw_withheld_authority.

..4+M [signer] M signer accounts.

Data expected by this instruction: WithdrawWithheldTokensFromMintData

WithdrawWithheldTokensFromAccounts = 2

Transfer all withheld tokens to an account. Signed by the mint’s withdraw withheld tokens authority. This instruction is susceptible to front-running. Use HarvestWithheldTokensToMint and WithdrawWithheldTokensFromMint as an alternative.

The withheld confidential tokens are aggregated directly into the destination available balance.

Note on front-running: This instruction requires a zero-knowledge proof verification instruction that is checked with respect to the account state (the currently withheld fees). Suppose that a withdraw withheld authority generates the WithdrawWithheldTokensFromAccounts instruction along with a corresponding zero-knowledge proof for a specified set of accounts, and submits it on chain. If the withheld fees at any of the specified accounts change before the WithdrawWithheldTokensFromAccounts is executed on chain, the zero-knowledge proof will not verify with respect to the new state, forcing the transaction to fail.

If front-running occurs, then users can look up the updated states of the accounts, generate a new zero-knowledge proof and try again. Alternatively, withdraw withheld authority can first move the withheld amount to the mint using HarvestWithheldTokensToMint and then move the withheld fees from mint to a specified destination account using WithdrawWithheldTokensFromMint.

In order for this instruction to be successfully processed, it must be accompanied by the VerifyWithdrawWithheldTokens instruction of the zk_elgamal_proof program in the same transaction or the address of a context state account for the proof must be provided.

Accounts expected by this instruction:

Single owner/delegate

[] The token mint. Must include the TransferFeeConfig extension.

[writable] The fee receiver account. Must include the TransferFeeAmount and ConfidentialTransferAccount extensions.

[] Instructions sysvar if VerifyCiphertextCiphertextEquality is included in the same transaction or context state account if VerifyCiphertextCiphertextEquality is pre-verified into a context state account.

[signer] The mint’s withdraw_withheld_authority.

..4+N [writable] The source accounts to withdraw from.

Multisignature owner/delegate

[] The token mint. Must include the TransferFeeConfig extension.

[writable] The fee receiver account. Must include the TransferFeeAmount and ConfidentialTransferAccount extensions.

[] Instructions sysvar if VerifyCiphertextCiphertextEquality is included in the same transaction or context state account if VerifyCiphertextCiphertextEquality is pre-verified into a context state account.

[] The mint’s multisig withdraw_withheld_authority.

..4+M [signer] M signer accounts.

5+M+1..5+M+N [writable] The source accounts to withdraw from.

Data expected by this instruction: WithdrawWithheldTokensFromAccountsData

HarvestWithheldTokensToMint = 3

Permissionless instruction to transfer all withheld confidential tokens to the mint.

Succeeds for frozen accounts.

Accounts provided should include both the TransferFeeAmount and ConfidentialTransferAccount extension. If not, the account is skipped.

Accounts expected by this instruction:

[writable] The mint.

..1+N [writable] The source accounts to harvest from.

Data expected by this instruction: None

EnableHarvestToMint = 4

Configure a confidential transfer fee mint to accept harvested confidential fees.

Accounts expected by this instruction:

Single owner/delegate

[writable] The token mint.

[signer] The confidential transfer fee authority.

*Multisignature owner/delegate 0. [writable] The token mint.

[] The confidential transfer fee multisig authority,

[signer] Required M signer accounts for the SPL Token Multisig account.

Data expected by this instruction: None

DisableHarvestToMint = 5

Configure a confidential transfer fee mint to reject any harvested confidential fees.

Accounts expected by this instruction:

Single owner/delegate

[writable] The token mint.

[signer] The confidential transfer fee authority.

*Multisignature owner/delegate 0. [writable] The token mint.

[] The confidential transfer fee multisig authority,

[signer] Required M signer accounts for the SPL Token Multisig account.

Data expected by this instruction: None