spl_token_2022_interface::extension

Trait Extension

pub trait Extension {
    const TYPE: ExtensionType;
}

Trait to be implemented by all extension states, specifying which extension and account type they are associated with

Required Associated Constants

const TYPE: ExtensionType

Associated extension type enum, checked at the start of TLV entries

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.